+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix new tabs opening from links
 - Fixed duplicate tab
+- Update downloads pop-up
 
 ## v0.11.1 - 2026-03-20